requirements.txt
# Flask Boilerplate - Dependencias
# ================================

Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Werkzeug==3.0.3
python-dotenv==1.1.0

# Producción
gunicorn==21.2.0
gevent>=24.2.1
gevent-websocket>=0.10.1

# Base de datos (comenta si usas SQLite local)
# psycopg2-binary>=2.9.11
